Only use BigInt at the data source. Added support for both Absolute & Relative fees

+++ lib/receive/domain/usecases/create_receive_swap_use_case.dart
@@ -24,7 +24,7 @@ class CreateReceiveSwapUseCase {
   Future<Swap> execute({
     required String walletId,
     required SwapType type,
-    required BigInt amountSat,
+    required int amountSat,
   }) async {
     try {
       final wallet = await _walletManager.getWallet(walletId);
